requirements.txt
apipkg==1.5
asgiref==3.2.3
cornac==1.2.2
Django==3.0
django-cors-headers==3.2.0
djangorestframework==3.10.3
execnet==1.7.1
gunicorn==20.0.4
joblib==0.14.1
numpy==1.17.4
pandas==0.25.3
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.22
scipy==1.3.3
six==1.13.0
sqlparse==0.3.0
tqdm==4.41.1
whitenoise==5.0.1
